pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zephyros"
version = "2.0.0"
authors = [
  { name="Keno Krieger", email="kriegerk@uni-bremen.de" },
]
description = "A package for wind power prediction"
readme = "README.md"
license = "GPL-3.0-only"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "pandas",
    "matplotlib",
    "scikit-learn",
    "tensorflow",
    "keras",
    "xgboost",
    "sklearn_rvm"
]

[project.urls]
homepage = "https://github.com/kenokrieger/zephyros"
repository = "https://github.com/kenokrieger/zephyros"
issues = "https://github.com/kenokrieger/zephyros/issues"
